Adding in source / position

+++ nonstop_wavfile_sink.h
@@ -80,7 +80,7 @@ public:
 	 */
 	virtual void set_bits_per_sample(int bits_per_sample) = 0;
 
-	virtual float length_in_seconds() = 0;
+	virtual double length_in_seconds() = 0;
 };
 
 } /* namespace blocks */
